Rename 'components/MonitoringItemsOverviewTableInfectionStatus.vue' to 'components/index/CardsMonitoring/MonitoringItemsOverview/InfectionStatus.vue'
