+++ components/index/CardsMonitoring/MonitoringItemsOverview/InfectionStatus.vue
